requirements.txt
# pip3 install -r requirements.txt

requests
dnspython
python-whois
python-nmap
nuclei